T4235: keep full paths in subtract tree
In the intial implementation, the delete tree only kept the first node
of a path not present in the RHS of the comparison; this was done as a
convenience for generating 'trimmed' delete commands for the CLI.
However, this loses needed information, and breaks the symmetry of
results:
        LHS, RHS -> del, inter, add
        RHS, LHS -> add, inter, del
Keep full paths in delete tree and rename trees.del to trees.sub(-tract).

A separate function 'trim_tree' will be provided to produce a
CLI-appopriate delete tree and commands.

(cherry picked from commit 7e4d9ff86ca96a990f52e05310a0486cb668cfb8)