T7605: add string sanitization functions for direct load of config file
It is necessary to escape single backslashes in config files before
parsing. In order to restore single/multiple backslashes on render, the
functions escape_backslash/unescape_backslash are used when
parsing/rendering from the Python side; as we also want to load/save
directly with vyconfd, add analogous functions to the library.